+++ charts.go
@@ -47,6 +47,7 @@ type ChartSelection struct {
 
 type ChartDataPageOutput struct {
 	Result     string
+	Locale     string
 	Type       string
 	AnalogData []PortData
 }
@@ -141,7 +142,7 @@ func getChartData(writer http.ResponseWriter, request *http.Request, params http
 		logError("CHARTS", "Problem loading timezone, setting Europe/Prague")
 		loc, _ = time.LoadLocation("Europe/Prague")
 	}
-	layout := "2006-01-02;15:04:05"
+	layout := "2006-01-02T15:04"
 	dateFrom, err := time.ParseInLocation(layout, data.From, loc)
 	dateFrom = dateFrom.In(time.UTC)
 	if err != nil {